+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Version 0.33
 Currently in development.
 
 - Improved search query builder
+- Added support for default values to object reference schemas
 
 
 Version 0.32